goschedstats: rename runnable.go to runnable_enabled.go
This naming makes things more clear, as `_enabled` and `_disabled` are
complementary in terms of build tags.

Epic: none
Release note: None
